fastapi-backend/pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.134.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
